feat: use modal to replace popover for UploadCart
